Move regression host module to regression integration test
