+++ tests/util/spectest.rs
@@ -208,9 +208,9 @@ impl GlobalValue for StaticGlobal {
     }
 }
 
-struct MutableStaticGlobal {
-    value: Mutex<Value>,
-    ty: ValType,
+pub struct MutableStaticGlobal {
+    pub value: Mutex<Value>,
+    pub ty: ValType,
 }
 
 impl GlobalValue for MutableStaticGlobal {
@@ -1011,78 +1011,6 @@ pub fn spectest_host_module() -> HostModule {
                 },
             ),
         }],
-    }
-}
-
-/// Extra host module used only by the regression integration tests.
-pub fn regression_host_module() -> HostModule {
-    HostModule {
-        name: "regression".into(),
-        globals: vec![
-            HostGlobal {
-                name: "mut_global_i32".into(),
-                value: spacewasm::Box::new(MutableStaticGlobal {
-                    value: Mutex::new(Value::I32(0)),
-                    ty: ValType::I32,
-                })
-                .unwrap()
-                .into_global_value_dyn(),
-            },
-            HostGlobal {
-                name: "mut_global_i64".into(),
-                value: spacewasm::Box::new(MutableStaticGlobal {
-                    value: Mutex::new(Value::I64(0)),
-                    ty: ValType::I64,
-                })
-                .unwrap()
-                .into_global_value_dyn(),
-            },
-            HostGlobal {
-                name: "mut_global_f32".into(),
-                value: spacewasm::Box::new(MutableStaticGlobal {
-                    value: Mutex::new(Value::F32(0.0)),
-                    ty: ValType::F32,
-                })
-                .unwrap()
-                .into_global_value_dyn(),
-            },
-            HostGlobal {
-                name: "mut_global_f64".into(),
-                value: spacewasm::Box::new(MutableStaticGlobal {
-                    value: Mutex::new(Value::F64(0.0)),
-                    ty: ValType::F64,
-                })
-                .unwrap()
-                .into_global_value_dyn(),
-            },
-        ],
-        functions: vec![
-            HostFunction::new(
-                "return_i32_from_all_args",
-                "iIfd".into(),
-                "i".into(),
-                |_, args| {
-                    let Value::I32(v) = args[0] else {
-                        unreachable!()
-                    };
-                    ControlFlow::Continue(Some(Value::I32(v)))
-                },
-            ),
-            HostFunction::new("return_i64", "".into(), "I".into(), |_, _| {
-                ControlFlow::Continue(Some(Value::I64(0x123456789)))
-            }),
-            HostFunction::new("return_f32", "".into(), "f".into(), |_, _| {
-                ControlFlow::Continue(Some(Value::F32(12.5)))
-            }),
-            HostFunction::new("return_f64", "".into(), "d".into(), |_, _| {
-                ControlFlow::Continue(Some(Value::F64(42.25)))
-            }),
-            HostFunction::new("noop", "".into(), "".into(), |_, _| {
-                ControlFlow::Continue(None)
-            }),
-        ],
-        memory: vec![],
-        table: vec![],
     }
 }
